requirements.txt
opencv-python==4.5.1.48
pandas==1.1.5
Pillow==8.2.0
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.1
torch==1.8.1
torchvision==0.9.1
tqdm==4.60.0
tokenizers==0.13.3
transformers==4.30.0